Consolidate world generation module
